# swagger/swagger.yaml
swagger: '2.0'
info:
  description: MonkeyRadio API Documentation
  version: 2.0.0
  title: MonkeyRadio API
  contact:
    email: contact@monkeyradio.fr
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - http
  - https
host: api.monkeyradio.fr
basePath: /api
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".
paths:
...
  /radio:
    post:
      summary: Create a new radio
      description: Require 'radio.create' permission
      security:
        - Bearer: []
      produces:
        - application/json
        - application/yaml
        - application/x-yaml
      tags:
        - Radio
      parameters:
        - in: body
          name: body
          description: radio object
          required: true
          schema:
            $ref: '#/definitions/radio'
      responses:
        '201':
          description: successful operation
        '400':
          description: Invalid request body
        '500':
          $ref: '#/definitions/InternalError'
    get:
      summary: Get all the radios
      description: Public route - Get all the radios informations including 'slug', 'name', 'icon' and 'live' (all radio variants slug and name used to the CDN)
      tags:
        - Radio
      produces:
        - application/json
        - application/yaml
        - application/x-yaml
      responses:
        '200':
          description: Radio List
          schema:
            type: array
            items:
              $ref: '#/definitions/radio'
        '400':
          description: Invalid status value
          schema:
            $ref: '#/definitions/InvalidResponse'
        '500':
          $ref: '#/definitions/InternalError'
    delete:
      summary: Delete a radio
      description: Require 'radio.delete' permission
      security:
        - Bearer: []
      tags:
        - Radio
      produces:
        - application/json
        - application/yaml
        - application/x-yaml
      parameters:
        - in: body
          name: body
          description: radio slug
          required: true
          schema:
            slug:
              type: string
              description: radio slug
              required: true
      responses:
        '200':
          description: successful deleted - Created
        '400':
          description: Invalid request body
        '500':
          $ref: '#/definitions/InternalError'
  /radio/{slug}:
    get:
      summary: Get the radio represented by slug
      description: Get the radio represented by slug
      tags:
        - Radio
      parameters:
        - name: slug
          in: path
          description: radio slug
          required: true
          type: string
      produces:
        - application/json
        - application/yaml
        - application/x-yaml
      responses:
        '200':
          description: Radio data
          schema:
            $ref: '#/definitions/radio'
        '400':
          description: This radio does not exist or this slug is not valid
          schema:
            $ref: '#/definitions/InvalidResponse'
        '500':
          $ref: '#/definitions/InternalError'
    put:
      summary: Update radio
      description: Require 'radio.update' permission
      security:
        - Bearer: []
      produces:
        - application/json
        - application/yaml
        - application/x-yaml
      tags:
        - Radio
      parameters:
        - name: slug
          in: path
          required: true
          type: string
        - in: body
          name: body
          description: radio object
          required: true
          properties:
            name:
              type: string
              description: radio name
            icon:
              type: string
              description: radio icon
      responses:
        '202':
          description: successful operation
        '400':
          description: Invalid request body
        '500':
          $ref: '#/definitions/InternalError'
...
definitions:
  radioVariant:
    type: object
    properties:
      slug:
        type: string
      name:
        type: string
      icon:
        type: string
  radio:
    type: object
    properties:
      slug:
        type: string
        description: radio slug
        required: true
      name:
        type: string
        description: radio name
      icon:
        type: string
        description: radio icon
      variants:
        type: array
        items:
          $ref: '#/definitions/radioVariant'
...
  InvalidResponse:
    type: object
    properties:
      message:
        type: string
  InternalError:
    description: Internal Server Error